+++ UI/Prefabs/Groups/S4Window.cs
@@ -26,7 +26,7 @@ namespace Forge.UX.UI.Prefabs.Groups {
 
         public override UIElement Instantiate() {
             ITextureCollectionManager tcm = DI.Dependencies.Resolve<ITextureCollectionManager>();
-            UIWindow element = new UIWindow(new NineSliceTextureComponent(tcm.Get((int)TextureCollectionMap.ForgeUI, (int)ForgeTextureMap.S4Window), 35));
+            UIWindow element = new UIWindow(new NineSliceTextureComponent(tcm.Get(TextureCollectionMap.ForgeUI, ForgeTextureMap.S4Window), 35));
             this.ApplyPropertyValues(element);
 
             InstantiateChildren(element);
